Complete an initial version of the RateLimiter code.
The primary goal of the rate limiter it to ensure the API's rate limit is not
exceeded. The secondary goal is for the sleep time between requests to be
relatively stable under the assumption that requests will continue to come in
immediately. This stability should occur both when there is only a single
reddit API client, and when there are multiple reddit API clients.

The decrease in the number of remaining API requests after a single request is
issued is an approximation for the number of reddit API clients. This
assumption appears to work reasonably well in my limited testing assuming all
clients follow the same procedure. However this approach does permit a greedy
reddit API client to starve all other clients, with clients using this
RateLimiter experiencing gradually longer delays between requests.

This approach should be satisfactory, however, given that the end user should
have control over all of their own reddit API clients.